rename die() to throw()
die() was a term modeled after a keyword in perl, but its use limited to a
single language makes it a lesser intuitive term.  On the other hand, throw()
is a better known term, but bashlib's offering does not offer the function that
is the typical counterpart to `throw`, `catch`, which may confuse some users.
With neither good option, throw feels like the easier read.