use the preferred form
Use '&>' instead of '>&' when redirecting both stdout and stderr.  Both are
equivalent but the former is the preferred form according to the bash
documentation.

For reasons that's not clear, this change requires quoting some unquoted
characters in the string.sh test code.  Not doing so causes the test to fail
if files named 1 and 2 exist in the current working directory.